archive/fetchhire/source_code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fetchhire"
version = "1.0.0"
description = "Advanced job scraping solution that bypasses 403 errors using Playwright headless browser technology"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ammr Abbasher", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Markup :: HTML",
]
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "snowflake-connector-python>=3.0.0",
    "pandas>=2.0.0",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/akabbas/fetchhire"
Repository = "https://github.com/akabbas/fetchhire"
Issues = "https://github.com/akabbas/fetchhire/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["scrapers*", "database*", "templates*", "tests*"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
